bump version, realtime point remapping, move semantics of analysis data, SOC in editor
this is a pretty big update. It revolves around allowing settings to change how the points in timbre space are mapped, from a more natural representation (which may have very significant clustering) to a uniform distribution. All other changes were in pursuit of this goal, but needed to be done anyway. These include incorporating move semantics for analysis data, so that when the timbre space representation is needed, it gets stolen by a subobject in the processor, and other calculations based on it are precomputed. The editor is now a value tree listener, and redraws the points as the mapping representation setting is adjusted. There is further separation of concerns in the drawing methods in editor. The processor's unnecessary destructor was removed.